tests: Write files to a temporary directory
t/tiff.t fails if run from a read-only location:

    t/tiff.t .................... convert: unable to open image `test.tif': Permission denied @ error/blob.c/OpenBlob/2924.
    Unable to open test.pdf for writing at /usr/share/perl5/vendor_perl/PDF/Builder/Basic/PDF/File.pm line 439.
    # Looks like your test exited with 13 just after 8.
    t/tiff.t .................... Dubious, test returned 13 (wstat 3328, 0xd00)
    Failed 5/13 subtests

That's because the test creates many temporary files. This patch fixes
it by placing the files into a temporary directory which is
automatically cleaned up on exit.

It will also help when runnig the tests in parallel. The tests won't
trip on the same-named files.